requirements.txt
Keras==2.4.3
Flask==1.1.1
opencv-python==4.2.0.34
tensorflow==2.2.0
keras-preprocessing==1.1.0
pillow==7.0.0
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.15.5
numpy>=1.9.2
scipy>=0.15.1
flask-cors==3.0.7